HudOperation: add a user_data field
The action callback handlers for the actions on the operation have (as
by normal convention) the GActionMap on which they are installed (ie:
the operation) as their user_data.

They probably also want to know about the context on which the HudAction
that created the action was installed (eg: the window or the
application).

Add a user_data field to store that and set it from the user_data given
when installing the HudActionEntries.