Gtk and Valaify the consumer demo
Start implementing GVTk (the GVariant toolkit) to convert GVariant
descriptions of UI into UIs.

We now construct a UI for the widgets from the demo (although the UI
description is currently hardcoded).

Actions are wired up.  It's now possible to do toggles and button
clicks in the consumer and see the result in the demo app.  Sliders are
not yet working.