HudActionPublisher: add support for multiple publishers
Each publisher must have an application ID specified -- either directly,
or by way of giving a GApplication.

The menu is now exported at a sensible path.

We now also produce (theoretical) signals when creating
HudActionPublisher to announce our presence to the service, including
the object path of the menu.

Update the demo program to the new API.