Fix Ubuntu UI Session callbacks.
On Touch Images where we use Ubuntu UI Session API to get
notifications when active window changes. The callbacks
are being invoked from a separate thread. We have to
pass the callbacks through g_idle_add() to process them
in main gmaincontext.

This also requires to change the type of AbstractWindow and
AbstractApp as we can't pass on the ubuntu_ui_session_properties
accross different threads.