Fix HUD service activation
The problem is that upstart sets the $DBUS_SESSION_BUS_ADDRESS environment variable to a different value that that found in $HOME/.dbus/session-bus/*, and does not include the variable in job environments. Therefore processes inside upstart's job environment have no idea where the the session bus is.

The main part of this fix grabs the DBus environment from $HOME/.cache/upstart/dbus-session"

The activation hack script is also improved to wait upon the service being started with a 10 second timeout, instead of just sleeping for 10 seconds.